move ftplugins to after/
This ensures that event when this plugin is loaded from a package, it
runs after the Vim distribution files.

Why?

Because, for example, Vim's Scheme ftplugin sets `b:undo_ftplugin` to a
fixed string, ignoring any pre-existing content. On the author's
machine, with this plugin installed as a package, the plugin files
executed first, causing `b:undo_ftplugin` to be overwritten by the Vim
ftplugin file. This leads to the mappings not being correctly torn down.